Complete exercise 4.21*: bloglist expansion, step9
Change the delete blog operation so that a blog can be deleted only by the user who added the blog.

+++ part4/controllers/blogs.js
@@ -47,6 +47,17 @@ blogsRouter.post('/', async (request, response, next) => {
  # Delete a blog by ID
  --------------------------------------------------*/
 blogsRouter.delete('/:id', async (request, response, next) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+
+  const blog = await Blog.findById(request.params.id)
+
+  if ( !blog.user || blog.user.toString() !== decodedToken.id.toString() ) {
+    return response.status(401).json({ error: 'blog must be created by same user for deletion' })
+  }
+
   try {
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
